+++ apps/test-app/frontend/package.json
@@ -43,9 +43,9 @@
     "@itwin/unified-selection": "workspace:*",
     "@itwin/unified-selection-react": "workspace:*",
     "@itwin/webgl-compatibility": "catalog:itwinjs-core-dev",
-    "@stratakit/bricks": "^0.1.0",
-    "@stratakit/foundations": "^0.1.0",
-    "@stratakit/icons": "^0.1.0",
+    "@stratakit/bricks": "catalog:stratakit",
+    "@stratakit/foundations": "catalog:stratakit",
+    "@stratakit/icons": "catalog:stratakit",
     "@opentelemetry/api": "^1.9.0",
     "@tanstack/react-table": "^8.21.2",
     "@test-app/common": "workspace:*",
