+++ packages/hierarchies-react/src/presentation-hierarchies-react.ts
@@ -14,12 +14,12 @@ export {
 export { useSelectionHandler } from "./presentation-hierarchies-react/UseSelectionHandler.js";
 export { HierarchyLevelDetails, useTree, useUnifiedSelectionTree } from "./presentation-hierarchies-react/UseTree.js";
 export { useIModelTree, useIModelUnifiedSelectionTree } from "./presentation-hierarchies-react/UseIModelTree.js";
-export { TreeNodeRenderer } from "./presentation-hierarchies-react/itwinui/TreeNodeRenderer.js";
-export { FilterAction } from "./presentation-hierarchies-react/itwinui/FilterAction.js";
-export { TreeRenderer } from "./presentation-hierarchies-react/itwinui/TreeRenderer.js";
-export { TreeErrorRenderer } from "./presentation-hierarchies-react/itwinui/TreeErrorRenderer.js";
-export { useFlatTreeNodeList, useErrorList, FlatTreeNode } from "./presentation-hierarchies-react/itwinui/FlatTreeNode.js";
-export { LocalizationContextProvider } from "./presentation-hierarchies-react/itwinui/LocalizationContext.js";
+export { TreeNodeRenderer } from "./presentation-hierarchies-react/stratakit/TreeNodeRenderer.js";
+export { FilterAction } from "./presentation-hierarchies-react/stratakit/FilterAction.js";
+export { TreeRenderer } from "./presentation-hierarchies-react/stratakit/TreeRenderer.js";
+export { TreeErrorRenderer } from "./presentation-hierarchies-react/stratakit/TreeErrorRenderer.js";
+export { useFlatTreeNodeList, useErrorList, FlatTreeNode } from "./presentation-hierarchies-react/stratakit/FlatTreeNode.js";
+export { LocalizationContextProvider } from "./presentation-hierarchies-react/stratakit/LocalizationContext.js";
 
 export { GenericInstanceFilter, HierarchyNode, HierarchyProvider, getLogger, setLogger } from "@itwin/presentation-hierarchies";
 export { SelectionStorage } from "@itwin/unified-selection";
