+++ packages/hierarchies-react/src/presentation-hierarchies-react/stratakit/FilterAction.tsx
