+++ packages/hierarchies-react/src/presentation-hierarchies-react/stratakit/FlatTreeNode.ts
