+++ packages/hierarchies-react/src/presentation-hierarchies-react/stratakit/RootErrorRenderer.tsx
