+++ packages/hierarchies-react/src/presentation-hierarchies-react/stratakit/TreeErrorRenderer.tsx
