+++ packages/hierarchies-react/src/presentation-hierarchies-react/stratakit/TreeNodeRenderer.tsx
