+++ packages/hierarchies-react/src/presentation-hierarchies-react/stratakit/TreeRenderer.tsx
